requirements.txt
autopep8==2.0.2
click==8.1.3
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
pycodestyle==2.10.0
Werkzeug==2.2.3
